+++ Runtime/Timer.cs
@@ -107,9 +107,10 @@ namespace Nonatomic.TimerKit
 		/// <param name="rangeEnd">The end of the range (lower value for TimeRemaining)</param>
 		/// <param name="interval">The interval at which to trigger callbacks</param>
 		/// <param name="callback">The callback to execute at each interval</param>
+		/// <param name="isRecurring">Whether this milestone should re-trigger every time the timer restarts</param>
 		/// <returns>The created TimerRangeMilestone</returns>
-		public virtual TimerRangeMilestone AddRangeMilestone(TimeType type, float rangeStart, float rangeEnd, float interval, Action callback) 
-			=> _timer.AddRangeMilestone(type, rangeStart, rangeEnd, interval, callback);
+		public virtual TimerRangeMilestone AddRangeMilestone(TimeType type, float rangeStart, float rangeEnd, float interval, Action callback, bool isRecurring = false) 
+			=> _timer.AddRangeMilestone(type, rangeStart, rangeEnd, interval, callback, isRecurring);
 		
 		/// <summary>
 		/// Removes a specific milestone from the timer.